+++ projects/dxt-lossless-transform-bc1/src/transforms/standard/mod.rs
@@ -92,45 +92,3 @@ pub(crate) unsafe fn transform_with_separate_pointers(
 pub(crate) unsafe fn untransform(input_ptr: *const u8, output_ptr: *mut u8, len: usize) {
     untransform::untransform(input_ptr, output_ptr, len);
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::test_prelude::*;
-    use crate::transforms::standard::transform;
-    use crate::transforms::standard::untransform::untransform;
-
-    #[test]
-    fn can_transform_untransform() {
-        for num_blocks in 1..=512 {
-            let input = generate_bc1_test_data(num_blocks);
-            let mut transformed = allocate_align_64(input.len()).unwrap();
-            let mut reconstructed = allocate_align_64(input.len()).unwrap();
-            let mut reference = allocate_align_64(input.len()).unwrap();
-
-            // Generate reference output
-            transform_with_reference_implementation(input.as_slice(), reference.as_mut_slice());
-
-            unsafe {
-                // Test transform
-                transform(input.as_ptr(), transformed.as_mut_ptr(), input.len());
-                assert_eq!(
-                    transformed.as_slice(),
-                    reference.as_slice(),
-                    "transform_bc1 produced different results than reference for {num_blocks} blocks"
-                );
-
-                // Test untransform
-                untransform(
-                    transformed.as_ptr(),
-                    reconstructed.as_mut_ptr(),
-                    transformed.len(),
-                );
-                assert_eq!(
-                    reconstructed.as_slice(),
-                    input.as_slice(),
-                    "untransform_bc1 failed to reconstruct original data for {num_blocks} blocks"
-                );
-            }
-        }
-    }
-}
