+++ projects/dxt-lossless-transform-bc1/src/transforms/standard/transform/portable32.rs
@@ -255,7 +255,7 @@ pub unsafe fn u32_unroll_8(input_ptr: *const u8, output_ptr: *mut u8, len: usize
 mod tests {
     use super::*;
     use crate::test_prelude::*;
-    use core::ptr::copy_nonoverlapping;
+    use crate::transforms::standard::untransform::untransform;
 
     type PermuteFn = unsafe fn(*const u8, *mut u8, usize);
 
@@ -264,73 +264,22 @@ mod tests {
     #[case(u32_unroll_2, "u32 unroll_2")]
     #[case(u32_unroll_4, "u32 unroll_4")]
     #[case(u32_unroll_8, "u32 unroll_8")]
-    fn test_portable32_aligned(#[case] permute_fn: PermuteFn, #[case] impl_name: &str) {
-        for num_blocks in 1..=512 {
-            let mut input = allocate_align_64(num_blocks * 8).unwrap();
-            let mut output_expected = allocate_align_64(input.len()).unwrap();
-            let mut output_test = allocate_align_64(input.len()).unwrap();
-
-            // Fill the input with test data
-            unsafe {
-                copy_nonoverlapping(
-                    generate_bc1_test_data(num_blocks).as_ptr(),
-                    input.as_mut_ptr(),
-                    input.len(),
-                );
-            }
-
-            transform_with_reference_implementation(
-                input.as_slice(),
-                output_expected.as_mut_slice(),
-            );
-
-            output_test.as_mut_slice().fill(0);
-            unsafe {
-                permute_fn(input.as_ptr(), output_test.as_mut_ptr(), input.len());
-            }
-
-            assert_implementation_matches_reference(
-                output_expected.as_slice(),
-                output_test.as_slice(),
-                &format!("{impl_name} (aligned)"),
-                num_blocks,
-            );
-        }
-    }
-
-    #[rstest]
-    #[case(u32, "u32 no-unroll")]
-    #[case(u32_unroll_2, "u32 unroll_2")]
-    #[case(u32_unroll_4, "u32 unroll_4")]
-    #[case(u32_unroll_8, "u32 unroll_8")]
-    fn test_portable32_unaligned(#[case] permute_fn: PermuteFn, #[case] impl_name: &str) {
+    fn portable32_transform_roundtrip(#[case] permute_fn: PermuteFn, #[case] impl_name: &str) {
         for num_blocks in 1..=512 {
             let input = generate_bc1_test_data(num_blocks);
+            let len = input.len();
+            let mut transformed = vec![0u8; len];
+            let mut reconstructed = vec![0u8; len];
 
-            // Add 1 extra byte at the beginning to create misaligned buffers
-            let mut input_unaligned = vec![0u8; input.len() + 1];
-            input_unaligned[1..].copy_from_slice(input.as_slice());
-
-            let mut output_expected = vec![0u8; input.len()];
-            let mut output_test = vec![0u8; input.len() + 1];
-
-            transform_with_reference_implementation(input.as_slice(), &mut output_expected);
-
-            output_test.as_mut_slice().fill(0);
             unsafe {
-                // Use pointers offset by 1 byte to create unaligned access
-                permute_fn(
-                    input_unaligned.as_ptr().add(1),
-                    output_test.as_mut_ptr().add(1),
-                    input.len(),
-                );
+                permute_fn(input.as_ptr(), transformed.as_mut_ptr(), len);
+                untransform(transformed.as_ptr(), reconstructed.as_mut_ptr(), len);
             }
 
-            assert_implementation_matches_reference(
-                output_expected.as_slice(),
-                &output_test[1..],
-                &format!("{impl_name} (unaligned)"),
-                num_blocks,
+            assert_eq!(
+                reconstructed.as_slice(),
+                input.as_slice(),
+                "Mismatch {impl_name} roundtrip for {num_blocks} blocks",
             );
         }
     }
