+++ projects/dxt-lossless-transform-bc1/src/transforms/standard/untransform/portable32.rs
@@ -253,7 +253,6 @@ pub unsafe fn u32_detransform_unroll_8(input_ptr: *const u8, output_ptr: *mut u8
 mod tests {
     use super::*;
     use crate::test_prelude::*;
-    use crate::transforms::standard::transform::u32;
 
     type DetransformFn = unsafe fn(*const u8, *mut u8, usize);
 
@@ -270,7 +269,7 @@ mod tests {
 
             unsafe {
                 // Transform using standard implementation
-                u32(original.as_ptr(), transformed.as_mut_ptr(), original.len());
+                transform(original.as_ptr(), transformed.as_mut_ptr(), original.len());
 
                 // Reconstruct using the implementation being tested
                 reconstructed.as_mut_slice().fill(0);
@@ -302,7 +301,7 @@ mod tests {
             // Transform using standard implementation
             let mut transformed = vec![0u8; original.len()];
             unsafe {
-                u32(original.as_ptr(), transformed.as_mut_ptr(), original.len());
+                transform(original.as_ptr(), transformed.as_mut_ptr(), original.len());
             }
 
             // Add 1 extra byte at the beginning to create misaligned buffers
